refactor forests and rivers and fix forests
