+++ pretty_gpx/city/data/rivers.py
@@ -1,5 +1,5 @@
 #!/usr/bin/python3
-"""Forest and rivers."""
+"""Rivers"""
 import os
 import numpy as np
 
@@ -19,30 +19,9 @@ from pretty_gpx.common.data.overpass_processing import get_polygons_from_relatio
 from pretty_gpx.common.data.overpass_processing import get_polygons_from_closed_ways
 
 
+
+
 RIVERS_CACHE = GpxDataCacheHandler(name='rivers', extension='.pkl')
-FORESTS_CACHE = GpxDataCacheHandler(name='forests', extension='.pkl')
-
-
-
-def download_city_forests(bounds: GpxBounds) -> Surface_Polygons:
-    """Download forest area from OpenStreetMap.
-
-    Args:
-        bounds: GPX bounds
-
-    Returns:
-        List of ShapelyPolygon of all forests/parks
-    """
-    cache_pkl = FORESTS_CACHE.get_path(bounds)
-
-    if os.path.isfile(cache_pkl):
-        forests_patches: Surface_Polygons = read_pickle(cache_pkl)
-    else:
-        forests_osm_results = _query_forests(bounds=bounds)
-        forests = get_polygons_from_relations(results=forests_osm_results)
-        forests_patches = create_patch_collection_from_polygons(forests)
-        write_pickle(cache_pkl, forests_patches)
-    return forests_patches
 
 
 def download_city_rivers(bounds: GpxBounds) -> Surface_Polygons:
@@ -68,16 +47,6 @@ def download_city_rivers(bounds: GpxBounds) -> Surface_Polygons:
         rivers_patches = create_patch_collection_from_polygons(rivers)
         write_pickle(cache_pkl, rivers_patches)
     return rivers_patches
-
-
-def _query_forests(bounds: GpxBounds):
-    """Query the overpass API to get the parks and forest of a city."""
-    result = overpass_query(['relation["leisure"="park"]["type"!="site"]',
-                             'relation["leisure"="park"]["type"!="site"]'],
-                            bounds,
-                            include_way_nodes=True,
-                            return_geometry=True)    
-    return result
 
 
 def _query_rivers_relations(bounds: GpxBounds):
